+++ restful-test-instances/organisations/src/main/java/org/restful/test/instances/service/OrganizationService.java
@@ -1,34 +1,22 @@
-/**
- * Copyright 2020 the project restful-test-instances authors
- * and the original author or authors annotated by {@author}
- * <p>
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * <p>
- * http://www.apache.org/licenses/LICENSE-2.0
- * <p>
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
 package org.restful.test.instances.service;
 
+import com.fasterxml.jackson.databind.JsonMappingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
 import lombok.experimental.FieldDefaults;
 import lombok.extern.slf4j.Slf4j;
-import org.modelmapper.ModelMapper;
 import org.restful.test.instances.model.detail.OrganizationDetail;
 import org.restful.test.instances.model.entity.Organization;
 import org.restful.test.instances.repository.OrganizationRepository;
+import org.restful.test.instances.service.specification.CustomSpecificationBuilder;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.Collections;
+import java.util.List;
 
-import static org.hibernate.validator.internal.util.Contracts.*;
+import static org.hibernate.validator.internal.util.Contracts.assertNotNull;
 
 
 /**
@@ -48,25 +36,108 @@ public final class OrganizationService {
 
     OrganizationRepository organizationRepository;
 
-    ModelMapper modelMapper;
+    ObjectMapper objectMapper;
 
-    public OrganizationDetail create(OrganizationDetail organizationDetailRequest) {
-        if (log.isInfoEnabled())
-            log.info("Получили запрос на создание сущности - {}", organizationDetailRequest);
+    public OrganizationDetail create(OrganizationDetail createOrganizationRequest) throws JsonMappingException {
+        log.info("Получили запрос на создание сущности - {}", createOrganizationRequest);
 
-        assertNotNull(organizationDetailRequest, "Организация не может быть null");
-        Organization organization = this.modelMapper.map(organizationDetailRequest, Organization.class);
-
-        if (log.isInfoEnabled())
-            log.info("Создали сущность организации - {}", organization);
+        assertNotNull(createOrganizationRequest, "Организация не может быть null");
+        Organization organization = this.objectMapper.updateValue(new Organization(), createOrganizationRequest);
+        log.info("Создали сущность организации - {}", organization);
 
         organization = this.organizationRepository.save(organization);
-        assertNotNull(organization.getId(), "Не получилось сохранить организацию");
-        OrganizationDetail organizationDetailResponse = this.modelMapper.map(organization, OrganizationDetail.class);
+        assertNotNull(organization.getUid(), "Не получилось сохранить организацию");
 
-        if (log.isInfoEnabled())
-            log.info("Возвращаем ответ - {}", organizationDetailResponse);
+        OrganizationDetail createOrganizationResponse = this.objectMapper.updateValue(
+                new OrganizationDetail(),
+                organization
+        );
+        log.info("Возвращаем ответ - {}", createOrganizationResponse);
 
-        return organizationDetailResponse;
+        return createOrganizationResponse;
+    }
+
+    public OrganizationDetail update(Long uidOrganization, OrganizationDetail updateOrganizationRequest)
+            throws JsonMappingException {
+        log.info(
+                "Получили запрос на обновлении сущности - {}\n с идентификатором - {}",
+                updateOrganizationRequest,
+                uidOrganization);
+
+        Organization organization = this.organizationRepository.findById(uidOrganization).orElse(null);
+        assertNotNull(
+                organization,
+                String.format("Организации с таким идентификатором - {} не существует", uidOrganization)
+        );
+        log.info("Получили сущность организации - {}", organization);
+
+        objectMapper.updateValue(organization, updateOrganizationRequest);
+        organizationRepository.save(organization);
+
+        OrganizationDetail updateOrganizationResponse = this.objectMapper.updateValue(
+                new OrganizationDetail(),
+                organization
+        );
+        log.info("Возвращаем ответ - {}", updateOrganizationResponse);
+
+        return updateOrganizationResponse;
+    }
+
+    public OrganizationDetail delete(Long uidOrganization) throws JsonMappingException {
+        log.info("Получили запрос на удаление сущности с идентификатором - {}", uidOrganization);
+
+        Organization organization = this.organizationRepository.findById(uidOrganization).orElse(null);
+        assertNotNull(
+                organization,
+                String.format("Организации с таким идентификатором - {} не существует", uidOrganization)
+        );
+        log.info("Получили сущность организации - {}", organization);
+
+        organizationRepository.delete(organization);
+
+        OrganizationDetail deleteOrganizationResponse = this.objectMapper.updateValue(
+                new OrganizationDetail(),
+                organization
+        );
+        log.info("Возвращаем ответ - {}", deleteOrganizationResponse);
+
+        return deleteOrganizationResponse;
+    }
+
+    public List<OrganizationDetail> find(List<Long> uid,
+                                   List<String> name,
+                                   List<String> inn,
+                                   List<String> kpp,
+                                   List<String> address) throws JsonMappingException {
+        log.info(
+                "Получили запрос на выборку сущностей:\nuid - {}\nname - {}\ninn - {}\nkpp - {}\naddress - {}",
+                uid,
+                name,
+                inn,
+                kpp,
+                address
+                );
+
+        List<Organization> organizations = this.organizationRepository.findAll(
+                CustomSpecificationBuilder.<Organization>getInstance()
+                        .withIn("uid", uid)
+                        .withIn("name", name)
+                        .withIn("inn", inn)
+                        .withIn("kpp", kpp)
+                        .withIn("address", address)
+                        .build()
+        );
+        log.info("Получили сущность организации - {}", organizations);
+
+        if (organizations == null) return Collections.emptyList();
+
+        List<OrganizationDetail> findOrganizationResponse = new ArrayList<>(){{
+            for (Organization organization : organizations) {
+                this.add(objectMapper.updateValue(new OrganizationDetail(), organization));
+            }
+        }};
+        log.info("Возвращаем ответ - {}", findOrganizationResponse);
+
+        return findOrganizationResponse;
     }
 }